gm_16F1: revert fn_8016F1F0 return type to short
The short→int change broke fn_8016C4F4 (100%→98.82%) because
the caller emits a sign-extend for the short return value that
matches the original binary.

+++ src/melee/gm/gm_16F1.c
@@ -97,7 +97,7 @@ Gm_DecType gmDecisionGetType(int kind)
     return curr->x5;
 }
 
-int fn_8016F1F0(int idx)
+short fn_8016F1F0(int idx)
 {
     return lbl_803D5A4C[idx].kind;
 }
